feat: implement useTimetable() hook
